+++ profile_config/__init__.py
@@ -19,7 +19,7 @@ from .merger import ConfigMerger
 from .profiles import ProfileResolver
 from .resolver import ProfileConfigResolver
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __all__ = [
     "ProfileConfigResolver",
     "ConfigDiscovery",